content/canvas/test/test_2d.imageData.create.tiny.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.create.tiny - bug 433004</title>
<!-- Testing: createImageData() works for sizes smaller than one pixel -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

var imgdata = ctx.createImageData(0.0001, 0.0001);
ok(imgdata.data.length == imgdata.width*imgdata.height*4, "imgdata.data.length == imgdata.width*imgdata.height*4");
ok(imgdata.width == 1, "imgdata.width == 1");
ok(imgdata.height == 1, "imgdata.height == 1");
var isTransparentBlack = true;
for (var i = 0; i < imgdata.data.length; ++i)
    if (imgdata.data[i] !== 0)
        isTransparentBlack = false;
ok(isTransparentBlack, "isTransparentBlack");

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
